Nasty compaction bug with OpenJDK classloader tables
In the OpenJDK class-library port, the classloader native
hashtables are hung off the private "classes" field in
java.lang.ClassLoader.  As this is a reference field, and
will be scanned by GC, the native pointer was stored in
a Long[1] array which was stored in the "classes" field.

On initialisation of a ClassLoader, the "classes" field
is initialised to an empty java.util.Hashtable.  To hijack
the field, we need to tell if we have already installed
a pointer to the C-based hashtable in it.  This was done
by checking if the type of the Object in "classes" was a
[J array.

This check fails in rare circumstances during compaction.
Specifically, if the [J array is earlier in the heap than
the ClassLoader object, when we reach the ClassLoader
object during phase 1 of compaction, the [J array has
already been threaded.  The C pointer in the [J array is
unaffected (as it's not a reference).  However, the
array's class pointer has been threaded.  The check of
the object's type therefore fails, and a NULL hashtable
pointer is returned.  This means the ClassLoader's hashtable
is not threaded, and the loaded class references are not
updated to the new addresses after compaction.

This is fixed by changing how the native C pointer is stored.
Instead of storing it in a [J array, the native pointer is
stored directly in the "classes" field.  Additionally, the
field is hidden from GC to prevent it being scanned.  This was
not previously possible when I first implemented the code.
Reference field hiding is needed for the "referent" field of
weak references, but the code was not factored out as a general
purpose hiding technique.  This was done later to support
sun.misc.Unsafe static field bases in the OpenJDK port.

Signed-off-by: Robert Lougher <rob@jamvm.org.uk>